added unit tests of uuid_mountpoint_factory

+++ lib/webfuse/adapter/impl/uuid_mountpoint_factory.c
@@ -7,20 +7,36 @@
 
 #include <stdlib.h>
 #include <string.h>
+#include <stdbool.h>
 
 struct wf_impl_uuid_mountpoint_factory
 {
     char * root_path;
+    bool root_created;
 };
 
 void *
 wf_impl_uuid_mountpoint_factory_create(
-    char * root_path)
+    char const * root_path)
 {
-	mkdir(root_path, 0755);
+    struct wf_impl_uuid_mountpoint_factory * factory = NULL;
+    bool root_created = false;
 
-    struct wf_impl_uuid_mountpoint_factory * factory = malloc(sizeof(struct wf_impl_uuid_mountpoint_factory));
-    factory->root_path = strdup(root_path);
+    struct stat info;
+    int rc = stat(root_path, &info);
+    if ((0 != rc) || (!S_ISDIR(info.st_mode)))
+    {
+        rc = mkdir(root_path, 0755);
+        root_created = true;
+    }
+
+    if (0 == rc)
+    {
+        factory = malloc(sizeof(struct wf_impl_uuid_mountpoint_factory));
+        factory->root_path = strdup(root_path);
+        factory->root_created = root_created;
+    }
+
 
     return factory;
 }
@@ -31,7 +47,11 @@ wf_impl_uuid_mountpoint_factory_dispose(
 {
     struct wf_impl_uuid_mountpoint_factory * factory = user_data;
 
- 	rmdir(factory->root_path);
+    if (factory->root_created)
+    {
+ 	    rmdir(factory->root_path);
+    }
+
     free(factory->root_path);
     free(factory);
 }
